Update patch set 8
Patch Set 8: Looks good to me, but someone else must approve

I failed - for now I've prevented swipes that would be invalid on the lockscreen.  In the future I'd like to remedy this, but it has a lot to do I think with how the lockscreen is laid out and I'd rather not mess with that too much right now.

Patch-set: 8