Update patch set 4
Patch Set 4: Looks good to me, but someone else must approve

It works good.  I like how when you go to select the rotary it disables the phone and messaging lockscreen so it doesn't get confused.  There are some whitespace issues in the code you could take care of.

Patch-set: 4