Update patch set 1
Patch Set 1:

The crash was when you actually went to open the Date and Time activity.  The crash originated in addPreferencesFromResource(R.xml.date_provider_settings)

I am not sure what in the xml caused it.  But the simple renaming of keys is what fixed it.  I checked for any kind of weird characters in the xml, and corruption of the xml itself.  Nothing turned up.

Patch-set: 1