Update patch set 1
Patch Set 1:

Thats fine, it's a simple merge to G.

I was trying to come up with a way to determine if we should disable any overclock on boot in case the user did something crazy and crashed the system. Maybe check the end of /proc/last_kmsg (if it exists) for badness. I'm getting ahead of myself now, it's just something that pissed me off about SetCPU a few times.

Patch-set: 1