Update patch set 8
Patch Set 8:

Any way to provide a setting in config.xml to select the default style in use for a freshly installed rom? For a LDPI device I'd like the old sliding tab style to be default. The rotary one is taking up way too much screen estate, and the new one can't even be moved further down the screen as it doesn't respect the marginBottom value anymore.

Patch-set: 8