Update patch set 8
Patch Set 8:

Sven: Click (Tattoo) and Buzz (Wildfire) are ldpi devices (and I have both of them lol). They use the mdpi framework though, as there's no official ldpi framework from Google. This works just fine most of the time, and only needs some custom changes here and there to make stuff fit and look better. Lockscreen is one example, and with all the stuff added to it lately, it's a bit of a puzzle making it not look like a total mess on the tiny ldpi screen. :) For example I moved the rotary slider all the way down to the bottom, as it otherwise would overlap with things above it. However, now it seems to be fixed in position, as my bottom margin value for it doesn't work anymore.

Steve: That sounds good. Thanks. :)

Patch-set: 8