Update patch set 8
Patch Set 8:

I just wanted to let you guys know that neither rotary or rotary revamped works on the Toshiba Folio tablet.

Apart from the fact that it's graphically cut off, no matter how long you try to drag it to unlock, it never registers it as enough.

The critical part is that rotary revamped is selected as default, which would have rendered the build unusable on the Folio unless menu-unlock worked by default as well.
An overlay (in case this issue isn't solved) is essential for the Folio to select another default lockscreen style.

Also the Lense style is graphically cut off which in effect only makes the default sliding tab style viable for the Folio.

Let me know if you need more info for this, I will try to take screenshots to post tonight.

Patch-set: 8