Update patch set 8
Patch Set 8:

Density is set to 120. I think a bottom margin of 20dip would be good with ldpi. It will overlap with the album art though, but not much that can be done with that atm, considering how the current lockscreen layout is designed.

Another thing I noticed, is that if I select lense style, it swithches to rotary style for the phone answer screen when someone calls. And when you answer the call, it stays on the screen behind the in-call buttons.

Patch-set: 8