Update patch set 2
Patch Set 2:

Do we want this option with the rest of the phone options in call settings or here?

I agree with the placement of the option here and think we should look into ditching/rewriting/revamping the current phone/contact options as they have been forward ported since donut.

Patch-set: 2