Update patch set 1
Patch Set 1:

I am not sure that this is the best way to do this. We have to use Samsung's closed source binary for sensors, and it is returning a bad maxRange for the light sensor (returns 3000 even though the max it actually returns is 15000). 

This adds a ro.sensors.light.maxrange that if present, is used instead of the value that the sensors lib returns. All current devices will use the sensors lib, Samsung SGS phones will have this set so that they can use this value until we are able to get all the sensors working using open source code.

Patch-set: 1