Update patch set 3
Patch Set 3:

Nebojsa, yea those were my original intentions.

However on irc, i was advised that it'd be better if other devices could also configure such a feature. Hence i kept it.

However if that is what will stand in the way of this being merged, i have no problem in putting a build.DEVICE check for galaxysmtd :)

Patch-set: 3