Update patch set 19
Patch Set 19:

Ok, I checked what exactly happens. Basically the issue is that the activity picker only offers ACTION_QUICK_CONTACT, but this one is only usable within the launcher, so we need ACTION_VIEW.

I will check out whether there's a better way to fix this, but for now, instead of sprinkling all those replaceAll's, we should have a common base function in some utility class in CMParts (e.g. 'String getPickResult(Intent data)' which returns the content URI) to have the replacement at a centralized place.
You can grep CMParts for PICK_ACTIVITY to see that this is needed for a lot of places ;-)

Patch-set: 19