Update patch set 19
Patch Set 19:

That'd be good :)
Given the amount of code duplication (you can e.g. grep for REQUEST_PICK_APPLICATION), I think that perhaps even a common base class ('AppPickingActivity') and deriving InputActivity, GestureCreateActivity etc. from it.

To finish this particular change, wouldn't it be worth a thought to just leave out the string replacement stuff, ignoring the contact crash problem for now and solve it in a different patch set? The 'long press menu' configuration stuff is not directly related to the crash, IMO...

Patch-set: 19