Update patch set 19
Patch Set 19:

Danny,

Ok so it happens because the sourceBounds are not set. If i set them it works fine but not what the user might want, it shows the quick contact view rather than taking into the contacts details. (quick contact im referring to the bar that slides above the contact)

One fix i could implement is that when launching quick contact, if a null value for sourcebounds is found, we could directly launch the contacts detail via an intent.

Let me know what you think of that.

Patch-set: 19