Update patch set 19
Patch Set 19: Looks good to me, but someone else must approve

Ok, so thinking this is good as it is. 

In regards to "sprinkling" all of those replaceAll's, until a fix is merged (for eg something like http://review.cyanogenmod.com/#change,6008), till then this will atleast prevent a fc without making changes too deep in the framework. (Note, ive modified my frameworks/base commit and removed entries of replaceAll from it).

Once that contacts commit is merged or a fix is found, the replaceAll's in cmparts activities can be revoked.

Patch-set: 19