Custom action on the double tap on the home key.
This modification allows user to specify any custom action on the double
tap action on the home key. Such functionality could be very useful on
devices with sensor buttons "Home" / "Menu" / "Search" / "Back", and a very
limited set of a real hardware buttons.
New setting is available under CyanogenMod Settings -> Input.

This functionality is disabled by default, and when it is disabled - it does not
changes system behavior in any little aspect.

This patch is only a front-end / configuratoin part of the functionality, the
functionality itself is implemented in the android_frameworks_base, corresponding
requst has been sent also.