Update patch set 4
Patch Set 4:

The problem you run into here is that the Email app at this point (the ProvisionParser) doesn't have any context of which email account it's dealing with.  All it knows is that the EasSyncService is syncing, but not which account it's syncing for.  Hence why I moved this from an account setting to a global setting.  

If I try to add this next to the SSL/TrustedSSL settings, it ends up making a HUGE mess of the interfaces and is literally 20+ class changes just to pass around this boolean to state whether you should or should not disable the policies...this is where I started.  Just look at how the SSL/TrustedSSL setting is passed around as a variable all over the place.

If I change this just to blindly ignore (not an option for the account, just a CM setting overall), this is roughly a global change...and adding the option would be nice to give to our end-users (with warning).

This app is a complete fscking mess.  Seriously!  I won't even go into the EAS protocol implementation.  BARF!  Just trying to add in a download count (ala iPhone) as opposed to just a date range, this will require some long hours trying to rev engineer the protocol and implementation.

I'll look again, but I doubt there is an easier and cleaner way than this.  I tried, I promise! :-/

Patch-set: 4