Change nightmode to 3700K and add a new all red render effect.
The idea here is to create a more useful nightmode
which uses a warmer display temperature ala f.lux
and redshift. This is easier on the eyes than
simply lowering brightness and is more useful
than an all red nightmode.

Also changes the arrays to contain resource
references instead of bare strings, and adds
proper labels for the calibrated renderfx widgets.

Change-Id: I9247e4188a823f8a0031e72b5ba889bbbc2c9ab4