Timeout for Security Lock (3/3)
Adds an option to CMParts to control the timeout when
the security lock should be enabled. Done with a Timer,
so the lockscreens switch even while the display is off.

This is the same as http://review.cyanogenmod.com/6526
and http://review.cyanogenmod.com/6027, but uses an
other approach.

I tried to change as less code as needed to get that
working. Many things are copied from the other
review-tickets. So credits also to Sergey Parshin and
Jonathan Vonk.

Change-Id: Iaaa495603c146cc48c7a89019bc84734a822f19e