Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

Your patchset has a confusing corner case when screen lock security is set to "None". It also lacks an overall toggle for the feature.

My corresponding patchset, #6527 (and the #6526 framework patch), implements this and explains the semantics to the user.

Patch-set: 1