Update patch set 1
Patch Set 1:

Danny,

Regarding the "add path to the variable".

Im not sure which is better. The current way, the processing/cropping is done once and saved as a compressed jpeg and in lockscreen is just read/set as the image. If i do it the variable way, lockscreen will have to do this processing each time ? 

Let me know if i misunderstood your idea of the implementation.

Patch-set: 1