Update patch set 7
Patch Set 7:

Since this uses a track list, I think it's not necessary to run through the complete installed packages. The downside of using a track list is that users cannot customize something has not yet notified. For instance, I can't customize a raining alert app in advance, because it has been sunny since this patch. A discovery routine using the old way for adding desired packages to track list, would be nice. I am not sure about per package customization. Is it possible to have multiple LED settings for a package? I have to disable LED for Android System package while in fact I just want to disable LED for the wifi available notification.

Patch-set: 7