Update patch set 4
Patch Set 4:

Yea,

Basically the view of the app needs to redrawn for this to take effect. This applies to all apps.

Would it be better i add "Needs restart to take effect" as summary ? Its technically not needed, but it might avoid any confusion as to it not working.

Patch-set: 4