Update patch set 1
Patch Set 1:

Pavel,

even if the end result is not visually appealing (as in your 'white' example, which is the same on the Defy, btw), why should we disable blending completely instead of letting the user try and see whether it's sufficient for him? Also, the larger problem, compared to the LED spacing, seems to be the lack of color levels to me. It's only 6 per color (making 216 colors in total), and not all of those levels are easily distinguishable by eye.
Having said that, I think a better idea, compared to just greying out the setting, would be to completely hide the option if desired in the overlay bool.

Patch-set: 1