Update patch set 7
Patch Set 7:

Forgot to mention in my last comment: I am aware that the behaviour of the LED code has optimization potential (e.g. using categories for grouping), but I want to get rid of the incredibly hacky current code and all of its shortcomings first and want to to defer behaviour improvements to (a) later patch(es).

Patch-set: 7