Update patch set 7
Patch Set 7:

Rick,

basically the problem is that you aren't supposed to set the color to none anymore (it's done via the LED mode now). It's still possible for you due to the LG star overlay. Given that I reevaluate the answer I gave to Steve earlier - we need to adapt the overlays prior to merging this patch. ;-)
Will prepare patches for the two offenders (Samsung C1 and LG Star).

Any other problems when _not_ setting the color to none?

Patch-set: 7