Update patch set 6
Patch Set 6:

For one, the resource used for overscroll is a drawable and not a simple color selection (which is why the framework patch includes a new png to be used for this). Second, I feel your method would require a complete rewrite of the Tmobile Theme Engine, as it would have to be able to select specific xml files, png's, .9's, and various color selections for said xml files.

Patch-set: 6