Update patch set 10
Patch Set 10: I would prefer that you didn't submit this

This is cool and a feature I'd like to see but while using my BT gps works it breaks the internal GPS.

I set it to internal in CMParts but when I turn the GPS on bluetooth is automaticly enabled but GPS is never turned on.

Somehow the internal gps settings isn't working as intended.

Tested on both ZTE Blade and ZTE V9/Light.

Patch-set: 10