Update patch set 10
Patch Set 10: No score

The gps providers are loaded at boot (standard framework). So at reboot I'm swapping my gps provider with the internal one. I didn't want to change too much in the framework, so only one hook-point to make it clean and easy. If you set your GPS back to internal, you need to reboot (to get the old provider loaded instead of mines). I agree that it is better to switch dynamic, but then I have way more code in the framework. Another issue is that the system identifies "GPS" as internal and uses it on multiple locations. 

So the only way to get this correctly is to replace the "GPS" named provider on change. Let me see if it can be done cleanly.

Patch-set: 10