Update patch set 10
Patch Set 10: No score

Confirmed. Internal GPS works again after a reboot.

Well, at least it should be informed, perhaps with a toast message that a reboot is needed for a change to take effect?

Other than that, great work, I find this a useful addition and something I think Android should have supported out-of-the box from the start.

Patch-set: 10