Update patch set 4
Patch Set 4:

I'm fine with doing the music category in a separate change.

As for the date format, it's a good idea to keep it consistent with the alarm. Given the DeskClock app would need to be modified for this, I'm not sure whether it's better to make the change or leave it as-is. I didn't know the alarm clock was formatted like this before.

Patch-set: 4