Update patch set 4
Patch Set 4:

Danny: Where do I select the calendar? It's possible I missed one of the commits since I'm unsure of which ones are the needed ones. I merged all of them I saw.

David: Can list the ones I need to merge so that I can test this? This is the first time it's been non-obvious.

Patch-set: 4