Update patch set 4
Patch Set 4:

Björn: For full functionality you'll need changes I1afe8288, I4d1dddcc, Ie9e2a370, and Iacca97ad. The first two are a basic calendar event implementation: next event on local calendar within 24 hours. The last two expand it to allow calendar and lookahead selection and whether to look just for events with a reminder.

The options are in CMParts > Lockscreen > Widget Options.

Danny: Set an alarm for a few minutes from now and then check the lockscreen. The day is always visible. Looking at where that value is set, it also appears that it's not using the current locale's time format.

Patch-set: 4