Update patch set 11
Patch Set 11:

I'm testing this on a N1 and I'm having an issue where settings are not saved if I press the home bottom after modifying them. 
e.g. change led color for the messagging app->press home button->new message notification still has the old led color and this is still shown as the choosen color in the led settings for the messagging app.
This is happening for all the app settings (color, blink rate...), not with the the advanced settings.

Patch-set: 11