Update patch set 12
Patch Set 12:

Tested but the save option doesn't work, nothing happens when touching it. Personally I would prefer an automatic saving process when exiting from the app settings(either pressing the back button or the home button) but a save option is fine too.  
In this case it would be nice to have a toast message confirming that the settings have been saved. I agree to move the note to the bottom of the package settings activity.

Patch-set: 12