Update patch set 15
Patch Set 15:

New patch set fixes the tab Kim pointed out and adds Eugenio's IT translation.

Eugenio, about the force mode: 'Force on' means 'Make the LED being lit even if the application didn't enable it'. 'Force on for event' is the same, but doesn't enable the LED for ongoing events (the no-clearable ones listed under 'Ongoing').

I'm open to all wording improvement suggestions. In fact, I'm not 100% happy with the current wording, but can't think of a better wording which still fits into the selection list.

Patch-set: 15