Update patch set 3
Patch Set 3:

The patch responce is an ill placed troll.  No need to be up in arms.  The confirmation box is indeed a good idea.  But I would make sure that the confirmation did spell out that this feature is illegal in some areas of distribution.  This would be one of the "are you really really sure" kind of deals.  Unfortunately, this feature can lead to severe concequences for some people.  (Thank you nanny state style government)

Patch-set: 3