Update patch set 3
Patch Set 3:

I have looked at it.  I've even pulled it and tested it.  Personally, it feels a little too easy to just move on through the confirmation.  Sorry to be hitting on this so hard, but keep in mind that this software project is distributed out of the united states where these laws can come down very hard on developers within the united states.  I'm not trying to push our laws on anyone.  Rather, I'm trying to protect the guys I chat with and work with on a regular basis.  And if that comes off as arrogance or pushiness, so be it.

The only change I would like to see is a second confirmation.  Something emphasizing the fact it is illegal in some areas.  This might seem excessive, but this excessiveness can be used in defense of the project if it ever needs to be.

Patch-set: 3