Update patch set 5
Patch Set 5:

I was unable to get a last_kmsg but below is a logcat from when it's rebooting. It seems it throws an exception when running the SQL command to get the calendar entries. 

http://pastebin.com/cwarUXJA

Note however that this didn't happen if I only selected my personal calendar. Then it showed up fine. To get it to happen I also added the "Svenska helgdagar" calendar provided by Google and set the look ahead setting to 3 days. Perhaps you can try adding that calendar to your account too and see if it breaks it.

Patch-set: 5