Update patch set 3
Patch Set 3: I would prefer that you didn't submit this

http://pastebin.com/BivgMRjq

Without setting anything up, selecting an application at random (GTalk in this instance) and selecting "Force LED on" and then saying "test this", the device will enter an endless bootloop. There is a flaw in the logic somewhere, as it's trying to parse a color (when it shouldn't be attempting to do this at all).

Patch-set: 3