Update patch set 18
Patch Set 18: I would prefer that you didn't submit this

Yes, that new patch does way too much (NB: Even if it were correct, it'd still be sloppy to remove the preference, but keep the translations).

The notice only should be removed if an RGB LED is not present. I don't think we need it for the dual color LED - I think it's safe to assume that e.g. an amber LED is always amber.
Ladios, do you have any objections to that, looking at the dual color LED mixing?

Patch-set: 18