Update patch set 20
Patch Set 20: I would prefer that you didn't submit this

Now that I thought about this a bit more, I realized that the 'Always pulse' option must not be removed, because it just controls whether the notification LED should also be lit when the screen is on. It is - despite the title, which we should leave alone because it's widely translated already - not related to actual pulsing of the LED.
Michael can you please remove the according parts of the patch (overlay bool and the changes in AdvancedActivity.java?

Patch-set: 20