Update patch set 1
Patch Set 1:

http://review.cyanogenmod.com/#change,7781 is also required.  You probably applied that too, since this change alone won't do anything but add the pref.

It should only go to the home screen if you release the button before a long press.  The recent apps will begin fading in with no delay, so you'll see it appear when going to the home screen.  This allows holding home and selecting an item on the recent apps list without waiting for the combined long-press and fade-in delay.  This is too common an operation to have that much of a delay.

(Having used this in practice for a couple weeks, I think it should also reduce the fade-in transition time when this is enabled.)

Patch-set: 1