Update patch set 17
Patch Set 17:

Tested on ZTE Light(v9) with Holux M-1000 bluetooth GPS

Good to see it no longer needs a reboot.
However, in GPS Test it shows the date as 11-09-05 (format is set to YY-MM-DD) but today is 11-09-22. With internal GPS it shows 11-09-22.

Other than that, it seems to work as it should on my ZTE Light.

BUT, I tried the same on my HTC Desire (bravo) and while I can pair the GPS it never connects when I enable the GPS in Android. Tried several reboots and powercycled the BT GPS a few times. As soon as I switched back to my ZTE Light it connected fine again but on the Desire it refused to work.

Patch-set: 17