Update patch set 5
Patch Set 5:

Karma down for Firefox which dropped my input. Gah. So again:

There's an overlay for display_settings.xml in devices/lge/star:
        <!-- Backlight, not supported on the star
        <PreferenceScreen android:key="backlight_settings"
            android:title="@string/backlight_title" android:summary="@string/backlight_summary">
            <intent android:action="android.intent.action.MAIN"
                android:targetPackage="com.cyanogenmod.cmparts"
                android:targetClass="com.cyanogenmod.cmparts.activities.BacklightActivity" />
        </PreferenceScreen>-->

So the description is clearly wrong for LG Start, but I tend to leave it unchanged and crate a separate change which allows an overlay to safely toggle whether backlight settings are supported or not via config.xml and adding a second description string for devices that do not support backlight. Copying display_settings.xml just seems wrong; it would have been updated for every other change that is made there.

Opinions?

Patch-set: 5