Default volume control media (2/2 CMParts)
This patch adds a preference to make media the default volume control
instead of ringer. It is good for tablet users who don't have phone
or users who don't want to change their ringer volume always. The
only problem is that no beep noise is made when changing volume for
media. If this is a problem just tell me and i'll (try) fix it.

Change-Id: If8f37d8fa80d230ba940f0a08f2479dfb59a11c6

Conflicts:

	res/values-es/strings.xml
	res/values/strings.xml
	res/xml/sound_settings.xml
	src/com/cyanogenmod/cmparts/activities/SoundActivity.java