Update patch set 8
Patch Set 8:

ok. i forgot to mention what is the symptom.
when you change the style between ring and others, custom apps summary is not updated right away. if you go down in menu and come back, or turn screen off and back on, it is fixed.
it's missing mCustomAppActivityPref.setSummary() at lines 419 and 425, or in updateStylePrefs(). if we add it in updateStylePrefs(), we have to make sure mCustomAppActivityPref is initialized in onCreate() before calling it.

Patch-set: 8