Update patch set 1
Patch Set 1:

Nice addition. It's not particularly what I would have implemented, though. What I particularly dislike is the fact that support for such high-level functionality is needed in NotificationManagerService.
My approach to this (if I just had more time...) would have been something like an '<All>' item in each category, which allows configuration of the whole category at once. This item would
- not have a category list
- have an additional 'Keep unchanged' item in each list

The values could be initialized from the single items: If they match, initialize accordingly, otherwise initialize to 'Keep unchanged'. When saving, all of the items would be overridden according to the selection.
What also would be a nice addition is changing the category assignment from a list in each package and a separate category create/delete activity to combined thing, just like the app group selection for the profiles.

What's your opinion on all that?

Patch-set: 1