Update patch set 1
Patch Set 1:

The downside of not doing it in NotificationManagerService would be that it would be impossible (or highly convoluted, such as setting up a Broadcast listener) to handle new "unconfigured" packages (since they won't be in the list and NotificationManagerService won't know about it). The main reason I'm doing this is to allow the default behavior for unconfigured packages to be customizable (e.g. to completely disable LED notification if one wishes to by setting the "unconfigured" category setting to Force Off). More work will also need to be done when you change a package to another category (as it will now not conform to the new category but rather the old until you do the mass edit again).

Patch-set: 1