Ringlock (2/3): Style options added to CM settings
This patch combines the two concurent lockscreen designs in an option:
http://review.cyanogenmod.com/8564 - http://review.cyanogenmod.com/8674
credits to Stefano Pignataro for the first and Prash for the second redesign

this patch also can be used to add even more styles in the future - if provied.
this patch also cleans up some coding.

Background Info: Lockscreen cannot be themed - so this is the only way except
tempering with framework-res.apk

Change-Id: Id2e5888057ad39a47bbe4e2fa9fcb3dc907b23b4