Update patch set 7
Patch Set 7:

The default color is read from the theme (it uses the horizontal progress bar style). Some themes, however, have progress bar styles that don't quite work. For example, the System theme has an orangish-yellow style for horizontal progress bars so it actually clashes with the rest of the status bar.

The alternative is to add a new battery bar style to the themes and let theme makers implement it at their own pace.

I'm good with either way.

Patch-set: 7