Update patch set 1
Patch Set 1:

That use case makes sense to me, but to achieve that, you force a disconnect between UI and actual behaviour: Unconfigured packages say that the mode is 'Use category settings', while, unless the category is configured, they're actually using the application's settings.

Using the application settings if the category is not configured definitely makes sense to me, so I think the best idea is making the 'Use category settings' the second item in the list and having some logic in the LED activity code that says 'if category is configured, use item 2 as default, else item 1'.

BTW, the translations of pref_trackball_force_mode_entries also need to be updated.

Patch-set: 1