Update patch set 1
Patch Set 1:

IMO, it makes sense the way it is right now. The setting for a package says "Use category settings" and if the user goes into the category setting page, he/she will see that it is set to "Use application settings".

Having the package setting jump between "Use category settings" and "Use application settings" would be confusing for the user. (Maybe the user will think that he/she will have to go to every package and set it to "Use category settings".)

As for the translation, what should I do? Obviously, I can't speak those languages. (Which is why we shouldn't be translating arrays but rather point those arrays to strings and translate those strings instead to not break things every time a value is added to that array.)

Patch-set: 1