Update patch set 3
Patch Set 3:

With more options being added to individual styles, I think there needs to be some separation between general settings that apply to all lockscreen styles (or picking the style itself) and settings that apply to only one (or a couple) styles.

That change is one of those changes that would benefit from only being shown when the Ring style is selected.

Patch-set: 3