Update patch set 3
Patch Set 3: Looks good to me, but someone else must approve

definitely a much cleaner look to it. easier for a user to figure out changes like this.

when merged with this: http://r.cyanogenmod.com/#change,8598 the "custom application" part is also updated as you change styles on-the-fly

Patch-set: 3