Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

Since this checkbox as well as the "show battery percentage" seem to be independent of one another, perhaps they should be combined into a multiple choice selection. Either that, or you need to find a way to remove the "small battery" icon next to the text when "show battery percentage" is enabled.

Patch-set: 1