Update patch set 7
Patch Set 7:

I do like the ability to pick the color if the theme is messing with the color.  As in my theme i have created, it does theme the progressbars so i have to choose a color.  But I probably do think a new style maybe the better route. 

I do have any idea though, how about making the user be able to pick whether to have the battery bar on top or bottom of the statusbar.  I think this would come in handy for tablets.

Patch-set: 7