Update patch set 4
Patch Set 4: I would prefer that you didn't submit this

We should push Android up, that it supports all policies and not ignore the policies. It should be good for enterprise, but this patch only produces hate by the exchange administrators. Maybe this could be the reason, why exchange administrators totally block android devices.

Patch-set: 4