Update patch set 2
Patch Set 2:

I was able to reproduce the issue.  Here is my output from the sql query:

1353|ringlock_style_pref|2
1357|rotary_style_pref|1
1380|in_call_style_pref|2
1381|lockscreen_style_pref|4

I have the lockscreen set as lens and the in-call set as rotary with the "normal style".  What I end up with on the lockscreen is this: http://dl.dropbox.com/u/6013923/device-2011-12-13-072109.png

Patch-set: 2