Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

(6 inline comments)

I'm also not convinced (at all) that this feature is needed in the userspace (as in: Java code) level. If a device seriously needs it, implement it in the device's liblights and add a device specific setting for it.

Besides that, there are a number of stilistic problems with this patch, see inline comments.

Patch-set: 1