Update patch set 1
Patch Set 1: No score

I think the main point was not to use spaces instead of hard tabs since the rest of the code did not use hard tabs. Don't feel bad about this we have a similar policy in my company. The benefit of using spaces is that it will look the same in all editors, but hard tabs may or may not look good in some editors.

Patch-set: 1