Update patch set 8
Patch Set 8: No score

As mentioned in the inline comments on patch set 3, popping up an EditText when "custom carrier label" is selected for the first time would be useful.

I can still implement this change, but can no longer test it. I lack a CM7 compatible device, though I'd be happy to forward-port this change to CM9 after this one is finished :)

Patch-set: 8