Update patch set 7
Patch Set 7:

I have been using this for several months, long before the lock volume keys patch was merged in. I have not had any issues with this patch. 

I enjoy having the volume keys control specifically the media instead of the ringer; I use the volume settings if I need to adjust any of the other volumes, which is extremely rare for me.

I can't speak for lock volume keys since I haven't used it because I've been a happy customer of this.

Patch-set: 7