Update patch set 5
Patch Set 5:

Works great on p990 & N1.

But on p970, it FC on CM Settings -> Display, as reported by my KANG users of p970. I don't own p970 but today I flashed a build with this cherry-picked on my colleague's p970 and it FC on display option. Later I rebuild w/o this picking and let my KANG users test it and the FC is gone. I am asking for logs but so far no one submitted. As such I can't confirm. The only confirmation is it FC on Display option on p970 device.

Patch-set: 5