Update patch set 1
Patch Set 1:

Glenn: I can understand that it can be frustrating but is has also vastly improved the code quality. I think the problem in this case was that there were issues mentioned that were not fixed or explained. I supposed it could also have used a few more reviewers too.

Patch-set: 1