Update patch set 2
Patch Set 2:

I've traced this all the way to PreferenceDialogFragment[1]. This patch simply restores its behavior to satisfy children classes which rely on it.
I'm aware of the problem in ApplicationLightPreference, an option is to simply refactor it to use the same logic as everything else. I already have that done locally, but it's broken further down the line (the colors don't stick when leaving and re-entering the activity, I suspect they're not being saved properly to CMSettings).

[1] https://github.com/CyanogenMod/android_frameworks_support/blob/cm-14.0/v14/preference/src/android/support/v14/preference/PreferenceDialogFragment.java#L253-L256

Patch-set: 2