Update patch set 2
Patch Set 2: Code-Review+1

When we set our custom OnClickListener in CustomPreferenceDialogFragment, OnClick() will not be called at all. As Dan said, however, onDialogClosed() needs onClick() to work properly.

And after some thinking I agree on this solution: we are (seemingly) "overriding" the OnClickListener so we should call "super" in the listener.

Patch-set: 2