Update patch set 1
Patch Set 1: Code-Review-1

> actually, wouldnt this completely overlap with the power menu? we
 > can't do this.

Yeah authorship needs to be assigned to Nate Benis but to your question, no doesn't overlay the power menu. While the lockscreen is turned off, you would longpress the power button and turn it on. To shut it off, you would press the power button once. While the screen is on, longpressing the power menu will still bring up the power menu.

Patch-set: 1