Update patch set 1
Patch Set 1:

> I thought we're going to allow free assignment of gestures to
 > actions?

Yeah, I was thinking of allowing users to map gesture.keycode to action (camera, torch, app, etc.) in the app. Then whenever the keyhandler receives a key event, it checks for a mapped gesture.keycode before starting the associated action.

When a gesture is not mapped by the user however, the gesture path is then disabled by calling setGestureEnabled().

Patch-set: 1