Update patch set 3
Patch Set 3:

> (1 comment)
 > 
 > Works fine.

since the new ROM was posted cm-14.1-20161212-NIGHTLY-osprey, my phone is having issues with com.android.phone.   I have tried clearing the cache and clearing the data for com.android.phone.   The error persists.  I also tried with the 1209, 8,10, and 11 Nightly with the same issue. Not able to make calls. Sim cards are not recognizing. Thanks

Patch-set: 3