Update patch set 1
Patch Set 1:

We should probably work on bringing this back. It is a nice feature, and there is quite a bit of space in the status bar header to house it (the weather info).

Patch-set: 1
Reviewer: Zhao Wei Liew <7880@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0