Update patch set 2
Patch Set 2:

Guys, I came up with this while trying to complement the following change: https://review.lineageos.org/#/c/57502/. I just put this as a draft to have your comments on first hand and to know if this could ever be accepted. This compiles and in theory it should work, but for some reason the broadcast receiver is not being registered. Let me know if you have any ideas how to accomplish this.

Patch-set: 2
Label: Verified=0