Update patch set 10
Patch Set 10: -Code-Review

> (1 comment)
 > 
 > I think this approach is inherently unsafe. When killing Settings
 > after the initial broadcast, the boolean value will be lost. Better
 > send a sticky broadcast and register the receiver in the fragment.
 > 
 > Besides those technical issues I still wonder whether we shouldn't
 > just nuke the 'fingerprint as home button' feature completely. It
 > somehow feels weird to me to begin with.

I appreciate your comments, but if you also feel that the feature should be nuked, it's pointless to loose more time with this change. I won't give up on the feature anyway and will do what's possible on the device side then. (for the record: I can change the action from there. The only thing I can't do is to disable the preference.)

Patch-set: 10